Three clippy reds and one suite red that came in with the loader's hang bound, fixed and filed
`bc4bb97e` landed with `cargo run -- --clippy` red on the bootloader — a
`u8::from` on a `u8` and two `format!`s with no argument — which blocks the gate
every worker folding onto this branch has to run. Mechanical, no semantics.

Its suite red is not mechanical and is filed rather than fixed, because it is
that landing's design question and not this branch's:
`issues/build/the-attempt-bound-refuses-a-second-qemu-boot-of-one-image.md`.
`is_the_retry` reads "the page holds nothing" as "the last boot reported
nothing", and under QEMU a healthy boot harvests nothing either — the RAM is
zeroed between launches — so the second launch of one image is a retry and boots
no kernel. `boot_partition_identity` boots one image more than once and is red
with `QEMU died before ===READY===`. Measured on `origin/metal-suite` alone in
the same session, red there too, twice; this branch changes nothing about it.

+++ bootloader/src/attempt.rs
@@ -67,7 +67,7 @@ pub fn read(system_table: &SystemTable<Boot>, guid: &[u8; 16]) -> Result<u8, Str
                 &buffer[..16]
             ));
         }
-        Ok(u8::from(buffer[16]))
+        Ok(buffer[16])
     })
     .and_then(|inner| inner)
 }
